pyproject.toml
[tool.poetry]
name = "project_manager"
version = "0.1.0"
description = ""
authors = ["Torokeldi <zhumagulovtorokeldi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
drf-spectacular = "^0.25.1"
django-debug-toolbar = "^3.8.1"
django-cors-headers = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
celery = "^5.2.7"
redis = "^4.5.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
ruff = "^0.0.247"
mypy = "^1.0.1"
black = "^23.1.0"
djangorestframework-stubs = "^1.8.0"
django-stubs = "^1.14.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = "tests.py test_*.py"
filterwarnings = ["ignore:SelectableGroups dict interface is deprecated. Use select.:DeprecationWarning"]

[tool.black]
line-length = 88
include = '\.pyi?$'

[tool.ruff]
exclude = ["migrations"]

[tool.isort]
profile = 'black'
line_length = 88
multi_line_output = 3
skip_glob = '**/migrations/*.py'
include_trailing_comma = true
float_to_top = true

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = true
exclude = ['venv']

[tool.django-stubs]
django_settings_module = "config.settings.local"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
